refactor(decomp): hoist BatchId + BlockPruneMode to foundation crates (Slice-D links 1-2)
Two zero-cascade leaf types move out of the root crate into foundation crates,
following the established re-export-shim pattern: the type definition moves
verbatim to the destination crate, and the old root path re-exports it so every
existing `crate::...` import keeps resolving (no big-bang reroute).

1. CompactBatchId  ->  crates/foundation/proximadb-kernel/src/batch_id.rs
   Old path src/storage/persistence/write_ahead_log/compact_batch_id.rs is now a
   one-line `pub use proximadb_kernel::batch_id::CompactBatchId;`. The WAL mod.rs
   `pub use compact_batch_id::CompactBatchId as BatchId;` alias resolves unchanged.
   The round-trip + uniqueness tests moved with the type (2 tests, both green in
   the kernel crate).

2. BlockPruneMode  ->  crates/foundation/proximadb-search-types/src/block_prune.rs
   The enum definition is removed from src/core/search/mod.rs and re-exported via
   the existing `pub use proximadb_search_types::{...}` block. Every
   `crate::core::search::BlockPruneMode` reference (sst/helix/swift engines,
   storage traits, tests) resolves unchanged.

Both types are pure leaf types (only std + serde deps; zero `use crate::` deps
verified by grep before the move). This is the first installments of the
clean-leaf extraction queue that unblocks the engine-traits port-inversion
chain, following the re-export pattern already proven by Slice D/D1.

StorageEngineType was scoped out of this PR: it is not a single leaf type but a
6-way duplicated enum (src/metrics, src/core/search/filter_contract,
src/storage/background_flush_context, src/storage/common/compaction_utils,
src/infrastructure/tier_data_movement, src/query/rl_planner/state) with
divergent variants, and several defs have `crate::proto::` impl-block cascade
deps. It is a consolidation/dedup task (cf. ROOT_CRATE_DECOMPOSITION_SLICE_D),
not a zero-cascade leaf hoist, so it is deferred.

Refs: docs/12-design/ROOT_CRATE_DECOMPOSITION_ENGINES_EXTRACTION_2026_07_12.adoc
       docs/12-design/ROOT_CRATE_DECOMPOSITION_CI_BUILD_OPTIMIZATION_2026_07_11.adoc

Verification:
- cargo check --lib --bins            (clean)
- cargo check --tests -p proximadb    (clean; shims resolve in test cfg)
- cargo check --tests -p proximadb-kernel / -p proximadb-search-types (clean)
- cargo nextest run -p proximadb-kernel         (138 pass; batch_id 2/2)
- cargo nextest run -p proximadb-search-types   (23 pass)
- cargo clippy --lib --bins           (exit 0, zero warnings)
- cargo fmt --all && cargo fmt --check (exit 0)
